refactor(helix-tui): split `widgets.rs` into submodules
