+++ helix-tui/src/layout.rs
@@ -12,3 +12,6 @@ pub use layout::Layout;
 
 mod alignment;
 pub use alignment::Alignment;
+
+mod corner;
+pub use corner::Corner;
